# backend/chatbot/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  description: |
    API cho hệ thống RAG Chatbot với xác thực JWT.
    
    ## Tính năng chính:
    - Xác thực và phân quyền người dùng
    - Quản lý người dùng
    - JWT token với access token và refresh token
    
    ## Bảo mật:
    - Sử dụng JWT Bearer token để xác thực
    - Access token có thời gian sống ngắn (2000s)
    - Refresh token có thời gian sống dài hơn (10000s)
  version: 0.0.1
  contact:
    name: Team 14
    email: team14@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /message/stream-create:
    post:
      tags:
        - Message
      summary: Tạo tin nhắn mới với streaming response
      description: Tạo một tin nhắn mới trong cuộc hội thoại và nhận phản hồi theo dạng streaming (Server-Sent Events)
      operationId: streamCreateMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
            examples:
              validMessage:
                summary: Tạo tin nhắn hợp lệ
                value:
                  text: "Xin chào, bạn có thể giúp tôi về AI không?"
                  conversationId: "conv-550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Streaming phản hồi thành công
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream với nội dung phản hồi
              example: "data: Xin\n\ndata: chào\n\ndata: !\n\n"
        '400':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Chưa xác thực hoặc token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /message/create:
    post:
      tags:
        - Message
      summary: Tạo tin nhắn mới
      description: Tạo một tin nhắn mới trong cuộc hội thoại và nhận phản hồi đầy đủ (yêu cầu xác thực)
      operationId: createMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
            examples:
              validMessage:
                summary: Tạo tin nhắn hợp lệ
                value:
                  text: "Xin chào, bạn có thể giúp tôi về AI không?"
                  conversationId: "conv-550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Tạo tin nhắn thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageResponse'
              examples:
                success:
                  summary: Tạo thành công
                  value:
                    code: 1000
                    message: "Success!"
                    data:
                      id: "msg-550e8400-e29b-41d4-a716-446655440000"
                      text: "Xin chào, bạn có thể giúp tôi về AI không?"
                      conversationId: "conv-550e8400-e29b-41d4-a716-446655440000"
                      userId: "550e8400-e29b-41d4-a716-446655440000"
                      isMe: true
        '400':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Chưa xác thực hoặc token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /message/list:
    get:
      tags:
        - Message
      summary: Lấy danh sách tin nhắn
      description: Lấy danh sách tin nhắn trong một cuộc hội thoại với phân trang (yêu cầu xác thực)
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: query
          required: true
          description: ID của cuộc hội thoại
          schema:
            type: string
          example: "conv-550e8400-e29b-41d4-a716-446655440000"
        - name: page
          in: query
          required: false
          description: Số trang (bắt đầu từ 1)
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: size
          in: query
          required: false
          description: Số lượng tin nhắn trên mỗi trang
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
      responses:
        '200':
          description: Lấy danh sách tin nhắn thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'
              examples:
                success:
                  summary: Danh sách tin nhắn với phân trang
                  value:
                    code: 1000
                    message: "Success!"
                    data:
                      currentPage: 1
                      pageSize: 10
                      totalPages: 2
                      totalElements: 15
                      data:
                        - id: "msg-1"
                          text: "Xin chào, bạn có thể giúp tôi về AI không?"
                          conversationId: "conv-550e8400-e29b-41d4-a716-446655440000"
                          userId: "550e8400-e29b-41d4-a716-446655440000"
                          isMe: true
                        - id: "msg-2"
                          text: "Chào bạn! Tôi có thể giúp gì cho bạn về AI?"
                          conversationId: "conv-550e8400-e29b-41d4-a716-446655440000"
                          userId: "bot-id"
                          isMe: false
        '401':
          description: Chưa xác thực hoặc token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Không tìm thấy cuộc hội thoại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    MessageRequest:
      type: object
      required:
        - text
        - conversationId
      properties:
        text:
          type: string
          description: Nội dung tin nhắn
          example: "Xin chào, bạn có thể giúp tôi về AI không?"
          minLength: 1
        conversationId:
          type: string
          description: ID của cuộc hội thoại chứa tin nhắn này
          example: "conv-550e8400-e29b-41d4-a716-446655440000"
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int64
          description: Mã trạng thái của phản hồi (1000 là thành công)
          example: 1000
          default: 1000
        message:
          type: string
          description: Thông điệp mô tả kết quả
          example: "Success!"
          default: "Success!"
        data:
          type: object
          description: Dữ liệu phản hồi (có thể là null)
          nullable: true
...
    MessageResponse:
      type: object
      properties:
        id:
          type: string
          description: ID duy nhất của tin nhắn
          example: "msg-550e8400-e29b-41d4-a716-446655440000"
        text:
          type: string
          description: Nội dung tin nhắn
          example: "Xin chào, bạn có thể giúp tôi về AI không?"
        conversationId:
          type: string
          description: ID của cuộc hội thoại chứa tin nhắn này
          example: "conv-550e8400-e29b-41d4-a716-446655440000"
        userId:
          type: string
          format: uuid
          description: ID của người dùng gửi tin nhắn
          example: "550e8400-e29b-41d4-a716-446655440000"
        isMe:
          type: boolean
          description: Tin nhắn có phải của người dùng hiện tại không
          example: true
    PageResponse:
      type: object
      properties:
        currentPage:
          type: integer
          description: Trang hiện tại (bắt đầu từ 1)
          example: 1
        pageSize:
          type: integer
          description: Số lượng phần tử trên mỗi trang
          example: 10
        totalPages:
          type: integer
          description: Tổng số trang
          example: 5
        totalElements:
          type: integer
          format: int64
          description: Tổng số phần tử
          example: 50
        data:
          type: array
          description: Dữ liệu của trang hiện tại
          items:
            type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int64
          description: Mã lỗi
          example: 1001
          enum:
            - 1001  # UNAUTHENTICATED
            - 1002  # USER_EXISTED
            - 1003  # USER_NOT_EXISTED
            - 1004  # ROLE_NOT_EXISTED
            - 1005  # UNAUTHORIZED
            - 1006  # UNCATEGORIZED
        message:
          type: string
          description: Thông điệp lỗi
          example: "authenticated error"
        data:
          type: object
          nullable: true
          description: Dữ liệu bổ sung về lỗi (thường là null)
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token để xác thực.
        
        **Cách sử dụng:**
        1. Đăng nhập qua `/auth/login` để lấy access token
        2. Thêm access token vào header: `Authorization: Bearer <token>`
        3. Access token có thời gian sống 2000 giây
        4. Khi hết hạn, sử dụng `/auth/refresh` với refresh token để lấy token mới
        
        **Lưu ý:**
        - Refresh token có thời gian sống 10000 giây
        - Token được ký bằng HS512 algorithm
security:
  - bearerAuth: []